refactor: Simplify form processing (steps one and two)
In order to reduce duplicate calls to Geonames API, removed state
field to process both domestic and international searches the same.
Return is now consistent and can be handled the same way each time.

User now enters only a city and country in step one and is returned
a results list in step two to select from.  If multiple results are
returned, they are now alphabetized by state/province for ease of
matching.